rating posts cleanup into different components
