+++ src/components/RatingsContainer/CommunityRatings/RatingsPosts/RatingPostsAvatar/RatingPostsAvatar.css
@@ -1,0 +1,25 @@
+.post-avatar {
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  border-radius: 1em;
+  padding: 0.1em 0.75em 0.1em 0.4em;
+  height: 2em;
+  text-decoration: none !important;
+  gap: 0.35em;
+}
+
+.post-avatar-img {
+  width: 1.8em;
+  height: 1.8em;
+  border-radius: 50%;
+}
+
+.post-avatar-img[src$=".svg"] {
+  height: 2em;
+}
+
+.post-avatar-name {
+  margin-left: 0.5em;
+  text-decoration: none;
+}
